adicionando controller e falta corrigir alguns erros

+++ src/TrayCorpChallenge.Test/Entities/ProductTests.cs
@@ -14,7 +14,7 @@ namespace TrayCorpChallenge.Test.Entities
         private  Product _product;
         private  string  _name = "Livro";
         private  bool _inventory = true;
-        private  double _value = 0;
+        private  decimal _value = 0;
 
 
         [TestMethod]
